Add files via upload
I checked the word document for each of the English sentences here (and often a sub-section of the sentence) and I found 3 sentences/phrases out of 45 that were not picked up by the script. I have inserted those 3 Burmese translations into the empty_burms.csv file. Also for the others not found, I have inserted a comment.