Re-written all the yaml files made missing ones
Rewritten all the yaml files. 
Not only were several missing completely but some were just in English because we didn't write out to yaml last time Lucy and I fixed the script. 
Now all of this is fixed. 

However, merging this will overwrite some of the manual edits that I made recently so I want to carefully check before merging.